Clear plotting device

Under some conditions, a device and its metadata need to be cleared manually. This can be done with either the new = TRUE argument within the call to Plot. Sometimes, the metadata of a previous plot will prevent correct plotting of a new Plot call. Use clearPlot to clear the device and all the associated metadata manually.

clearPlot(
  dev = dev.cur(),
  removeData = TRUE,
  force = FALSE,
  verbose = getOption("quickPlot.verbose")
)

# S4 method for numeric,logical
clearPlot(
  dev = dev.cur(),
  removeData = TRUE,
  force = FALSE,
  verbose = getOption("quickPlot.verbose")
)

# S4 method for numeric,missing
clearPlot(dev, force)

# S4 method for missing,logical
clearPlot(removeData, force)

# S4 method for missing,missing
clearPlot(dev, removeData, force)

Arguments

dev

Numeric. Device number to clear.

removeData

Logical indicating whether any data that was stored in the .quickPlotEnv should also be removed; i.e., not just the plot window wiped.

force

Logical or "all". Sometimes the graphics state cannot be fixed by a simple clearPlot(). If TRUE, this will close the device and reopen the same device number. If "all", then all quickPlot related data from all devices will be cleared, in addition to device closing and reopening.

verbose

Numeric or logical. If TRUE or >0, then messages will be shown. If FALSE or 0, most messages will be suppressed.

See also

Author

Eliot McIntire

Examples

if (interactive()) {
  Plot(1:10)
  clearPlot() # clears
}