Specify where to plot

Switch to an existing plot device, or if not already open, launch a new graphics device based on operating system used. On Windows and Mac, if no x is provided, then this will open or switch to the first non R Studio device, which is much faster than the png-based R Studio Plot device. Currently, this will not open anything new

dev(x, ...)

Arguments

x
The number of a plot device. If missing, will open a new non-RStudio plotting device

...
Additional arguments passed to newPlot.

Value

Opens a new plot device on the screen. Invisibly returns the device number selected.

Details

For example, dev(6) switches the active plot device to device #6. If it doesn't exist, it opens it. NOTE: if devices 1-5 don't exist they will be opened too.

Examples

# NOT RUN {
  dev(4)
# }