Convert plotArgs to list of lists

Internal function. Take the inputs as plotArgs to the Plot function, and make them a list of length numQuickPlotObjects entries of lists.

.makeList(plotArgs, numQuickPlotObjects)

# S4 method for list
.makeList(plotArgs, numQuickPlotObjects)

Arguments

plotArgs
list. Any arguments that the the grid package can accept for the specific grob types, e.g., rasterGrob, polygonGrob, etc.

numQuickPlotObjects
Numeric. The number of .quickPlotObjects. This can't easily be deduced from the plotArgs because of the RasterStacks. So passed manually.