Open a new plotting window

Open a new plotting window

newPlot(noRStudioGD = TRUE, ..., verbose = getOption("quickPlot.verbose"))

dev.useRSGD(useRSGD = FALSE)

Arguments

noRStudioGD

Logical Passed to dev.new. Default is TRUE to avoid using RStudio graphics device, which is slow.

...

Additional arguments.

verbose

Numeric or logical. If TRUE or >0, then messages will be shown. If FALSE or 0, most messages will be suppressed.

useRSGD

Logical indicating whether the default device should be the RStudio graphic device, or the platform default (quartz on macOS; windows on Windows; x11 on others, e.g., Linux).

Note

dev.new() is supposed to be the correct way to open a new window in a platform-generic way; however, does not work in RStudio (SpaDES#116). Use dev.useRSGD(FALSE) to avoid RStudio for the remainder of this session, and dev.useRSGD(TRUE) to use the RStudio graphics device. (This sets the default device via the device option.)

See also

Author

Eliot McIntire and Alex Chubaty

Examples

if (FALSE) {
  ## set option to avoid using Rstudio graphics device
  dev.useRSGD(FALSE)

  ## open new plotting window
  newPlot()
}