Parse arguments and find environments

Internal function used within objectNames.

.parseArgs(y, e, eminus1)

Arguments

y

A character representation of the arguments passed to a function, e.g., Plot.

e

Environment in which the function (e.g., Plot) was called.

eminus1

The parent environment of e.

Value

A list of length 2, with names objs and envs giving the standardized representation (i.e., replacing [[]] with $ notation for objects) of objects and their layers (if RasterStacks).