Prepare raster for plotting

Internal function. Takes a raster .quickPlotGrob, and converts zoomExtent into a zoom, and legendRange into a legend. Then calculates the maxpixels to plot for speed.

.prepareRaster(
  grobToPlot,
  zoomExtent,
  legendRange,
  takeFromPlotObj,
  arr,
  speedup,
  newArr
)

Arguments

grobToPlot

A .quickPlotGrob object.

zoomExtent

an extent object

legendRange

a numeric vector of length >=2 indicating the desired legend range.

takeFromPlotObj

logical. Should the object be found in the Plot call or .GlobalEnv

arr

an .arrangement object

speedup

numeric, greater than 1 will usually speed up plotting at the expense of resolution

newArr

logical, whether this is a new arrangement or just adding to a previous one

Author

Eliot McIntire