The .quickPlot class

This class contains all necessary information to build a Plot on a device, except the actual data. Thus, this class differs notably from the grid package, which keeps a copy of all data and information in a hidden location for further access for rebuilding, erasing etc. This difference allows the Plot function to be much faster than using the grid methodology directly. The cost to this speed gain is that the objects must be available, by name, in the .GlobalEnv.

Details

This class contains two slots, one for the overall arrangement of the plots within the device window, and the second for all the .quickPlotGrob() objects within that device window. These .quickPlotGrob() objects are the individual "smallest" plot unit.

These gp* parameters will specify plot parameters that are available with gpar(). gp will adjust plot parameters, gpText will adjust title and legend text, gpAxis will adjust the axes. size adjusts point size in a SpatialPoints object. These will persist with the original Plot call for each individual object. Multiple entries can be used, but they must be named list elements and they must match the ... items to plot. This is true for a RasterStack also, i.e., the list of named elements must be the same length as the number of layers being plotted. The naming convention used is: RasterStackName$layerName, i.e, landscape$DEM.

Slots

arr

An .arrangement object

quickPlotGrobList

list. A list of lists of .quickPlotGrob objects

See also

Author

Eliot McIntire