quickPlot

quickPlot uses S4 classes. "Dot" classes are not exported and are therefore intended for internal use only.

Plotting classes - used within <code>Plot</code>

New classes

.arrangement

The layout or "arrangement" of plot objects

.quickPlot

Main class for Plot - contains .quickGrob and .arrangement objects

.quickPlotGrob

GRaphical OBject used by quickPlot - smallest unit

Unions of existing classes:

.quickPlottables

The union of all object classes Plot can accept

.quickPlotObjects

The union of spatialObjects and several others

spatialObjects

The union of several spatial classes

See also

Plot