The quickPlot package

A high-level plotting system, built using 'grid' graphics, that is optimized for speed and modularity. This has great utility for quick visualizations when testing code, with the key benefit that visualizations are updated independently of one another.

Note

The suggested package fastshp can be installed with install.packages("fastshp", repos = "http://rforge.net", type = "source").

See also

Useful links:

http://quickplot.predictiveecology.org

https://github.com/PredictiveEcology/quickPlot

Report bugs at https://github.com/PredictiveEcology/quickPlot/issues