Plot spatial grobs (using grid package)

Internal function. Plot a raster Grob, a points Grob, polygon Grob.

.plotGrob(grobToPlot, col = NULL, real = FALSE, size = unit(5, "points"),
  minv, maxv, legend = TRUE, legendText = NULL, length = NULL,
  gp = gpar(), gpText = gpar(), pch = 19, speedup = 1,
  name = character(), vp = list(), ...)

# S4 method for SpatialPoints
.plotGrob(grobToPlot, col = NULL, real = FALSE,
  size = unit(5, "points"), minv, maxv, legend = TRUE, legendText = NULL,
  length = NULL, gp = gpar(), gpText = gpar(), pch = 19, speedup = 1,
  name = character(), vp = list(), ...)

# S4 method for matrix
.plotGrob(grobToPlot, col = NULL, real = FALSE,
  size = unit(5, "points"), minv, maxv, legend = TRUE, legendText = NULL,
  length = NULL, gp = gpar(), gpText = gpar(), pch = 19, speedup = 1,
  name = character(), vp = list(), ...)

# S4 method for SpatialPolygons
.plotGrob(grobToPlot, col = NULL, real = FALSE,
  size = unit(5, "points"), minv, maxv, legend = TRUE, legendText = NULL,
  length = NULL, gp = gpar(), gpText = gpar(), pch = 19, speedup = 1,
  name = character(), vp = list(), ...)

# S4 method for SpatialLines
.plotGrob(grobToPlot, col = NULL, real = FALSE,
  size = unit(5, "points"), minv, maxv, legend = TRUE, legendText = NULL,
  length = NULL, gp = gpar(), gpText = gpar(), pch = 19, speedup = 1,
  name = character(), vp = list(), ...)

Arguments

grobToPlot
Raster*, SpatialLines*, SpatialPoints*, or SpatialPolygons* object.

col
Currently only used for the legend of a Raster* object.

real
Logical indicating whether the data are real numbers (i.e., as opposed to integer or factor).

size
The size of the SpatialPoints.

minv
The minimum value on a Raster*. Required because not all Rasters have this defined internally.

maxv
The maximum value on a Raster*. Required because not all Rasters have this defined internally.

legend
Logical indicating whether a legend should be drawn. Default TRUE.

legendText
Vector of values to use for legend value labels. Defaults to NULL which results in a pretty numeric representation. If Raster* has a Raster Attribute Table (rat; see raster package), this will be used by default. Currently, only a single vector is accepted.

length
Numeric.

gp
grid parameters, usually the output of a call to gpar.

gpText
gpar object for legend label text.

pch
Point character for SpatialPoints, as par.

speedup
Numeric. The factor by which the number of vertices in SpatialPolygons and SpatialLines* will be subsampled. The vertices are already subsampled by default to make plotting faster.

name
Character string of name of object being plotted.

vp
whole viewport tree of quickPlotGrob

...
Additional arguments. None currently implemented.

Details

speedup is only used for SpatialPolygons, SpatialPoints, and SpatialLines in this function. Attempts have been made to subsample at a good level that optimizes speed of plotting, without losing visible quality. Nevertheless, to force all points to be plotted, use a speedup value less than 0.1. From a speed perspective, there appears to be an optimal subsampling when using thin from the fastshp package. Presumably, too much thinning requires large distance matrices to be calculated, slowing plotting down. Too little thinning causes an overabundance of points to be plotted, slowing plotting down.

The suggested package fastshp can be installed with: install.packages("fastshp", repos = "http://rforge.net", type = "source").

NOTE: you may get errors relating to not having installed the software tools required for building R packages on your system. For building on Windows, you'll need to install Rtools from https://cran.r-project.org/bin/windows/Rtools/.