The .quickPlotGrob class

This class contains the plotting .quickPlotGrob information.

Details

These gp* parameters will specify plot parameters that are available with gpar(). gp will adjust plot parameters, gpText will adjust title and legend text, gpAxis will adjust the axes. size adjusts point size in a SpatialPoints object. These will persist with the original Plot call for each individual object. Multiple entries can be used, but they must be named list elements and they must match the ... items to plot. This is true for a RasterStack also, i.e., the list of named elements must be the same length as the number of layers being plotted. The naming convention used is: RasterStackName$layerName, i.e, landscape$DEM.

Slots

plotName

character. Name of the plot frame, which is by default a concatenation of the objName and layerName

objName

character. Name of object represented by this .quickPlotGrob

envir

environment. The environment in which to find the objName

layerName

character. Name of the layer represented by this .quickPlotGrob. Primarily used for RasterStacks.

objClass

character. Class of the object represented by this .quickPlotGrob.

isSpatialObjects

logical. TRUE if the object is one of the SpaDES recognized spatialObject classes.

plotArgs

list. Any parameters needed for plotting, set by Plot call.

See also

quickPlotClasses